Move Semaphore generator
