+++ lib/generators/rolemodel/ci/semaphore/README.md
