+++ lib/generators/rolemodel/ci/semaphore/templates/heroku-deployment-commands.sh
