+++ lib/generators/rolemodel/ci/semaphore/templates/production-deploy.yml.erb
