+++ lib/generators/rolemodel/ci/semaphore/templates/semaphore.yml.erb
