+++ lib/generators/rolemodel/ci/semaphore/templates/staging-deploy.yml.erb
